Move fetch apis and onLoadContent to domUpdates.js
