+++ src/scripts.js
@@ -1,7 +1,8 @@
 import $ from 'jquery';
-import users from './data/users-data';
-import recipeData from  './data/recipe-data';
-import ingredientData from './data/ingredient-data';
+// import users from './data/users-data';
+// import recipeData from  './data/recipe-data';
+// import ingredientData from './data/ingredient-data';
+import domUpdates from './domUpdates.js'
 import User from './user';
 import Recipe from './recipe';
 import './css/base.scss';
@@ -13,30 +14,30 @@ let newRecipeObjects;
 
 let userID = Math.floor((Math.random() * 50) + 1);
 
-function getUsers() {
-  const userDataApi = "https:fe-apps.herokuapp.com/api/v1/whats-cookin/1911/users/wcUsersData"
-  const promise = fetch(userDataApi)
-    .then (response => response.json())
-    // .then(users => users.wcUsersData.find((user) => user.id === userID))
-  return promise;
-}
-
-function getRecipes() {
-  const recipeDataApi = "https://fe-apps.herokuapp.com/api/v1/whats-cookin/1911/recipes/recipeData"
-  const promise = fetch(recipeDataApi)
-    .then(response => response.json())
-    // .then((recipe) => pickles = recipe.recipeData)
-    // .then(() => console.log(createRecipeObject(pickles)))
-  return promise;
-}
-
-function getIngredients() {
-  const ingredientDataApi = "https://fe-apps.herokuapp.com/api/v1/whats-cookin/1911/ingredients/ingredientsData"
-  const promise = fetch(ingredientDataApi)
-    .then(response => response.json())
-    // .then((ingredient) => ingredient.ingredientsData
-  return promise;
-}
+// function getUsers() {
+//   const userDataApi = "https:fe-apps.herokuapp.com/api/v1/whats-cookin/1911/users/wcUsersData"
+//   const promise = fetch(userDataApi)
+//     .then (response => response.json())
+//     // .then(users => users.wcUsersData.find((user) => user.id === userID))
+//   return promise;
+// }
+//
+// function getRecipes() {
+//   const recipeDataApi = "https://fe-apps.herokuapp.com/api/v1/whats-cookin/1911/recipes/recipeData"
+//   const promise = fetch(recipeDataApi)
+//     .then(response => response.json())
+//     // .then((recipe) => pickles = recipe.recipeData)
+//     // .then(() => console.log(createRecipeObject(pickles)))
+//   return promise;
+// }
+//
+// function getIngredients() {
+//   const ingredientDataApi = "https://fe-apps.herokuapp.com/api/v1/whats-cookin/1911/ingredients/ingredientsData"
+//   const promise = fetch(ingredientDataApi)
+//     .then(response => response.json())
+//     // .then((ingredient) => ingredient.ingredientsData
+//   return promise;
+// }
 
 // Promise.all([userDataApi , ingredientDataApi])
 //   .then((values) => {
@@ -76,15 +77,19 @@ searchBtn.addEventListener("click", searchRecipes);
 showPantryRecipes.addEventListener("click", findCheckedPantryBoxes);
 searchForm.addEventListener("submit", pressEnterSearch);
 
-function onLoadContent() {
-  let promise1 = Pantry.getIngredients()
-  let promise2 = Recipe.getRecipes()
-  let promise3 = User.getUsers(userId)
-  Promise.all([promise1, promise2, promise3])
-    .then(values => {
-      //DOM shenanigans
-    })
-}
+// function onLoadContent() {
+//   let promise1 = Pantry.getIngredients()
+//   let promise2 = Recipe.getRecipes()
+//   let promise3 = User.getUsers(userId)
+//   Promise.all([promise1, promise2, promise3])
+//     .then(values => {
+//       // .then((ingredient) => ingredient.ingredientsData
+//       // .then((recipe) => pickles = recipe.recipeData)
+//       // .then(users => users.wcUsersData.find((user) => user.id === userID))
+//
+//       //DOM shenanigans
+//     })
+// }
 
 // GENERATE A USER ON LOAD
 function generateUser() {
